Conflicts:
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/.bowerrc
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/.gitignore
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/LICENSE
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/package.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/PuppetJs/src/puppet.js
	src/BuildSystem/Polyjuice/StaticFiles/sys/core-component-page/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/core-component-page/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/core-component-page/bowager-logo.png
	src/BuildSystem/Polyjuice/StaticFiles/sys/core-component-page/demo.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/core-component-page/index.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/LICENSE.txt
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/src/json-patch-duplex.js
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/src/json-patch-duplex.ts
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/src/json-patch.js
	src/BuildSystem/Polyjuice/StaticFiles/sys/fast-json-patch/src/json-patch.ts
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/LICENSE
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/gruntfile.js
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/imported-template.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/imported-template/package.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/index.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/package.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/juicy-redirect/src/juicy-redirect.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/model-operations/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/model-operations/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/model-operations/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/model-operations/model-operations.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/polymer/layout.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/polymer/polymer.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/.bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/index.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/package.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-js/src/puppet-js.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-redirect/README.md
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-redirect/bower.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-redirect/index.html
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-redirect/package.json
	src/BuildSystem/Polyjuice/StaticFiles/sys/puppet-redirect/src/puppet-redirect.html
	src/BuildSystem/WhiteList.txt
	src/Level1.sln
	src/Samples/LoadAndLatency/run_loadandlatency_test.bat
	src/Samples/PolePosition/run_poleposition_test.bat
	src/Samples/QueryProcessingTest/MetadataTest.cs
	src/Samples/QueryProcessingTest/run_queryprocessingtest.bat
	src/Samples/QueryProcessingTest/secondrun.bat
	src/Samples/SQLTest/run_sqlregression_test.bat
	src/Samples/SqlCacheTrasher/run_sqlcachetrasher_test.bat
	src/Starcounter.Apps.JsonPatch/AppRestServer.cs
	src/Starcounter.Internal/Handle.cs
	src/Starcounter.Node/Node.cs
	src/Starcounter.Rest/UriHandlersManager.cs
	src/Starcounter.Rest/UriManagedHandlersCodegen.cs
	src/Starcounter.Web.Tests/TestHandlerLevels.cs
	src/Starcounter.XSON.Tests/PuppetTests.cs
	src/Starcounter/Metadata/RuntimeView.cs
	src/Starcounter/Starcounter.csproj
	src/scnetworkgateway/scripts/gw_test_client_run.bat
	src/scnetworkgateway/scripts/gw_test_pair_run.bat
	src/scnetworkgateway/scripts/gw_test_server_run.bat

+++ src/Starcounter.Installer/Starcounter.InstallerEngine/InstallerMain.cs
@@ -69,7 +69,7 @@ namespace Starcounter.InstallerEngine
         }
         
         // Creates server config XML.
-        public static void CreateServerConfig(
+        public static void PointToServerConfig(
             String serverName,
             String pathToServerDir,
             String whereToSaveConfig)
@@ -90,13 +90,17 @@ namespace Starcounter.InstallerEngine
         // Copy gateway config XML.
         public static void CopyGatewayConfig(
             String serverDir,
-            String internalSystemPort)
+            String internalSystemPort,
+            Boolean useExistingServerInstallation)
         {
             // Copying gateway configuration.
-            File.Copy(
-                Path.Combine(InstallerMain.InstallationDir, ConstantsBank.ScGatewayConfigName),
-                Path.Combine(serverDir, ConstantsBank.ScGatewayConfigName),
-                true);
+            if (!useExistingServerInstallation) {
+
+                File.Copy(
+                    Path.Combine(InstallerMain.InstallationDir, ConstantsBank.ScGatewayConfigName),
+                    Path.Combine(serverDir, ConstantsBank.ScGatewayConfigName),
+                    true);
+            }
 
             // Overwriting server config values.
             if (!Utilities.ReplaceXMLParameterInFile(
@@ -694,6 +698,9 @@ namespace Starcounter.InstallerEngine
                 }
                 catch (InstallerAbortedException userCanceled)
                 {
+                    // Rolling back the installation.
+                    UninstallEngine.RollbackInstallation();
+
                     // Simply re-throwing user all cancellations.
                     throw userCanceled;
                 }
@@ -780,6 +787,9 @@ namespace Starcounter.InstallerEngine
                 }
                 catch (InstallerAbortedException userCanceled)
                 {
+                    // Rolling back the installation.
+                    UninstallEngine.RollbackInstallation();
+
                     // Simply re-throwing all user cancellations.
                     throw userCanceled;
                 }
